rename: global 폴더 이동
