+++ src/main/java/com/example/codionbe/global/common/CustomException.java
@@ -1,6 +1,6 @@
-package global.common;
+package com.example.codionbe.global.common;
 
-import global.common.exception.ErrorCode;
+import com.example.codionbe.global.common.exception.ErrorCode;
 import org.springframework.http.HttpStatus;
 
 public class CustomException extends RuntimeException {
