+++ src/main/java/com/example/codionbe/global/common/exception/ErrorCode.java
@@ -1,8 +1,8 @@
-package global.common.success;
+package com.example.codionbe.global.common.exception;
 
 import org.springframework.http.HttpStatus;
 
-public interface SuccessCode {
+public interface ErrorCode {
     HttpStatus getStatus();
     String getCode();
     String getMessage();
