+++ src/main/java/com/example/codionbe/global/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package global.config;
+package com.example.codionbe.global.config;
 
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Info;
